Species Dropdown Selection bug
This addresses an issue where any variant of Hybrid is selected  when creating or editing a flower. When you save the product and then return to edit it, the species field doesn't stay populated.

I changed the "paper-item key" to match the values that it should be looking for. Please check it out and advise on whether or not this is an acceptable fix and that it won't break a bunch of stuff.